Remove snapshots, comment, omit two small redundancies.

+++ gadjid/src/lib.rs
@@ -163,7 +163,7 @@ pub(crate) mod test {
             .collect();
 
         // parents_of returns a slice, (defined .iter() order), so we don't need to stabilize with sort.
-        let pa_true_1st_T = g_true.parents_of(ts[0] as usize).to_vec();
+        let pa_true_1st_T = g_true.parents_of(ts[0]).to_vec();
 
         // below, we sort results because the order of the elements in the hashsets is not defined and we want fully matching snapshots
         let mut an_true_1st_T: Vec<usize> = graph_operations::ancestors(&g_true, [ts[0]].iter())
